Move block crypt functionality to a new file
